terrarium/dl_gitmanager/pyproject.toml

[tool.poetry]
name = "datalens-gitmanager"
version = "0.0.1"
description = ""
authors = ["DataLens Team <datalens-opensource@yandex-team.ru>"]
packages = [{include = "dl_gitmanager"}]
license = "Apache 2.0"
readme = "README.md"


[tool.poetry.dependencies]
attrs = ">=22.2.0"
python = ">=3.10, <3.12"
gitpython = ">=3.1.37"
datalens-cli-tools = {path = "../dl_cli_tools"}

[tool.poetry.group.tests.dependencies]
pytest = ">=7.2.2"
[build-system]
build-backend = "poetry.core.masonry.api"
requires = [
  "poetry-core",
]

[tool.poetry.scripts]
dl-git = "dl_gitmanager.scripts.gitmanager_cli:main"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra"
testpaths = []

[datalens_ci]
skip_test = true

[tool.mypy]
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
strict_optional = true
